requirements.txt
torch
torchvision
numpy
pillow
scikit-image
opencv-python
streamlit
tqdm
kagglehub
matplotlib
